lib: avoid mixing `XmlBigCount` and `size_t`
This avoids some assumptions that these types are the same, or that the
range of one subsumes the range of the other. It also makes some
overflow checks more intuitive.

+++ expat/lib/internal.h
@@ -173,8 +173,8 @@ void _INTERNAL_trim_to_complete_utf8_characters(const char *from,
                                                 const char **fromLimRef);
 
 #if defined(XML_GE) && XML_GE == 1
-unsigned long long testingAccountingGetCountBytesDirect(XML_Parser parser);
-unsigned long long testingAccountingGetCountBytesIndirect(XML_Parser parser);
+size_t testingAccountingGetCountBytesDirect(XML_Parser parser);
+size_t testingAccountingGetCountBytesIndirect(XML_Parser parser);
 const char *unsignedCharToPrintable(unsigned char c);
 #endif
